requirements.txt
streamlit
PyMuPDF
PyPDF2
yake
keybert
sentence-transformers
networkx
matplotlib
whisper
torch
faiss-cpu
graphviz
spacy
scikit-learn
numpy
python-docx
python-pptx
pandas
regex
